Build fixes for linux

+++ src/dynd/dim_iter.cpp
@@ -72,7 +72,7 @@ void dynd::make_strided_dim_iter(
     out_di->data_elcount = 0;
     out_di->data_stride = stride;
     out_di->flags = dim_iter_restartable | dim_iter_seekable;
-    if (tp.get_data_size() == stride) {
+    if ((intptr_t)tp.get_data_size() == stride) {
         out_di->flags |= dim_iter_contiguous;
     }
     out_di->eltype = ndt::type(tp).release();
@@ -219,7 +219,7 @@ void dynd::make_buffered_strided_dim_iter(
     out_di->data_elcount = 0;
     out_di->data_stride = buffer_stride;
     out_di->flags = dim_iter_restartable | dim_iter_seekable;
-    if (buf.get_dtype().get_data_size() == buffer_stride) {
+    if ((intptr_t)buf.get_dtype().get_data_size() == buffer_stride) {
         out_di->flags |= dim_iter_contiguous;
     }
     out_di->eltype = ndt::type(val_tp).release();
